Added a cancel button to register page
